feat: refactor so that YouTube IFrame API can be enabled from configuration or action
